:recycle: Reworking delegation gen macro and making it available for #[bundle]
