+++ crates/bolt-lang/attribute/delegate/Cargo.toml
@@ -12,6 +12,4 @@ edition = { workspace = true }
 proc-macro = true
 
 [dependencies]
-syn = { workspace = true }
-quote = { workspace = true }
-proc-macro2 = { workspace = true }
+bolt-attribute.workspace = true
